Talk to kolibri-daemon in kolibri-gnome
