+++ src/kolibri-daemon.in
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-@PYTHON@ -m "kolibri_gnome.kolibri_service.main"
+@PYTHON@ -m "kolibri_gnome.kolibri_daemon.main"
