+++ src/kolibri_gnome/kolibri_daemon/kolibri_service_stop.py
