--- src/kolibri_gnome/kolibri_service/utils.py
@@ -1,10 +1,0 @@
-from contextlib import contextmanager
-
-
-@contextmanager
-def gapplication_hold(application):
-    application.hold()
-    try:
-        yield
-    finally:
-        application.release()
