+++ src/kolibri_gnome/meson.build
@@ -2,10 +2,12 @@ python = import('python')
 python_installation = python.find_installation('python3')
 
 kolibri_gnome_sources = [
+    'dbus_utils.py',
     'globals.py',
+    '__init__.py',
+    'kolibri_daemon_proxy.py',
     'kolibri_globals.py',
     'kolibri_settings.py',
-    '__init__.py',
 ]
 
 python_installation.install_sources(
@@ -25,29 +27,27 @@ python_installation.install_sources(
     subdir: 'kolibri_gnome/desktop_launcher'
 )
 
-kolibri_service_sources = [
-    'kolibri_service/application.py',
-    'kolibri_service/content_extensions.py',
-    'kolibri_service/__init__.py',
-    'kolibri_service/kolibri_service_main.py',
-    'kolibri_service/kolibri_service_monitor.py',
-    'kolibri_service/kolibri_service.py',
-    'kolibri_service/kolibri_service_setup.py',
-    'kolibri_service/kolibri_service_stop.py',
-    'kolibri_service/main.py',
-    'kolibri_service/utils.py',
+kolibri_daemon_sources = [
+    'kolibri_daemon/application.py',
+    'kolibri_daemon/content_extensions.py',
+    'kolibri_daemon/__init__.py',
+    'kolibri_daemon/kolibri_service_main.py',
+    'kolibri_daemon/kolibri_service_monitor.py',
+    'kolibri_daemon/kolibri_service.py',
+    'kolibri_daemon/kolibri_service_setup.py',
+    'kolibri_daemon/kolibri_service_stop.py',
+    'kolibri_daemon/main.py',
 ]
 
 python_installation.install_sources(
-    kolibri_service_sources,
-    subdir: 'kolibri_gnome/kolibri_service'
+    kolibri_daemon_sources,
+    subdir: 'kolibri_gnome/kolibri_daemon'
 )
 
 search_provider_sources = [
     'search_provider/application.py',
     'search_provider/__init__.py',
     'search_provider/main.py',
-    'search_provider/utils.py',
 ]
 
 python_installation.install_sources(
