--- src/kolibri_gnome/search_provider/utils.py
@@ -1,10 +1,0 @@
-from contextlib import contextmanager
-
-
-@contextmanager
-def gapplication_hold(application):
-    application.hold()
-    try:
-        yield
-    finally:
-        application.release()
