+++ src/meson.build
@@ -1,8 +1,8 @@
 subdir('kolibri_gnome')
 
 configure_file(
-    input: 'kolibri-daemon.in',
-    output: 'kolibri-daemon',
+    input: 'kolibri-gnome.in',
+    output: 'kolibri-gnome',
     configuration: kolibri_gnome_config,
     install: true,
     install_dir: bindir,
@@ -10,11 +10,11 @@ configure_file(
 )
 
 configure_file(
-    input: 'kolibri-gnome.in',
-    output: 'kolibri-gnome',
+    input: 'kolibri-daemon.in',
+    output: 'kolibri-daemon',
     configuration: kolibri_gnome_config,
     install: true,
-    install_dir: bindir,
+    install_dir: kolibri_gnome_libexecdir,
     install_mode: 'rwxr-xr-x'
 )
